requirements.txt
pillow
pycryptodome
keyring
python-dotenv
sqlalchemy
argon2-cffi
pyperclip
watchdog
PyQt6
